Added packaging system
- Can be pip installed
- Versioning determined automatically
- Can be run using python3 -m VMEC2GK
