+++ VMEC2GK/profile_plotter.py
@@ -9,7 +9,7 @@ import numpy as np
 import netCDF4 as nc
 import matplotlib.pyplot as plt
 
-from utils import (
+from .utils import (
     parse_input_file,
     ifft_routine,
     extract_essence,
