+++ pyproject.toml
@@ -1,0 +1,12 @@
+[build-system]
+requires = [
+    "setuptools>=61.0",
+    "setuptools_scm[toml] >= 6.2",
+    "setuptools_scm_git_archive",
+    "wheel >= 0.29.0",
+]
+build-backend = "setuptools.build_meta"
+
+[tool.setuptools_scm]
+write_to = "VMEC2GK/_version.py"
+git_describe_command = "git describe --dirty --tags --long --match v* --first-parent"
