+++ setup.cfg
@@ -1,3 +1,49 @@
+[metadata]
+
+name = VMEC2GK
+description = VMEC to gyrokinetics interface for tokamaks
+long_description = file: README.md
+long_description_content_type = text/markdown
+url = https://github.com/rahulgaur104/VMEC2GK
+author = Rahul Gaur
+#license = TODO
+#license_file = LICENSE.md
+keywords = plasma, physics, tokamak, research, gyrokinetics
+classifiers =
+    Programming Language :: Python
+    Development Status :: 3 - Alpha
+    Natural Language :: English
+    Intended Audience :: Science/Research
+    #License :: OSI Approved :: TODO
+    Operating System :: OS Independent
+    Programming Language :: Python
+    Topic :: Software Development :: Libraries :: Python Modules
+    Topic :: Scientific/Engineering :: Physics
+project_urls =
+    Source = https://github.com/rahulgaur104/VMEC2GK
+    Tracker = https://github.com/rahulgaur104/VMEC2GK/issues
+
+[options]
+packages = find:
+include_package_data = True
+python_requires = >= 3.6
+install_requires =
+    numpy ~= 1.20.3
+    scipy ~= 1.6.3
+    netCDF4 ~= 1.5.6
+    matplotlib ~= 3.4.2
+    xarray ~= 0.10
+    importlib-metadata; python_version >= "3.6"
+
+[options.extras_require]
+dev = 
+    flake8
+    black
+    build
+    twine
+tests = 
+    pytest >= 3.3.0
+
 [flake8]
 max-line-length = 88
 extend-ignore = E203
