--- src/__init__.py
@@ -1,10 +1,0 @@
-from .profile_plotter import plot_profiles
-from .eikcoefs_final import vmec_to_bishop
-
-# TODO
-# eikcoefs final
-# bishoper gs2
-# bishoper gx
-# bishoper ball
-
-__all__ = ["plot_profiles", "vmec_to_bishop"]
